Estructura para subtareas y cambio de color de app

+++ app/src/main/java/com/juan/notes/ui/notesDetail/NoteDetailFragment.kt
@@ -18,7 +18,7 @@ class NoteDetailFragment : Fragment() {
     private val noteViewModel: NoteDetailViewModel by viewModels()
 
     private lateinit var note: Note
-    private var noteId: Int = 0
+    private var noteId: Long = 0
 
     private var editMode: Boolean = false
     private var deleteNote: Boolean = false
@@ -28,9 +28,9 @@ class NoteDetailFragment : Fragment() {
         super.onCreate(savedInstanceState)
         setHasOptionsMenu(true)
         arguments?.let {
-            noteId = it.getInt("id_note")
+            noteId = it.getLong("id_note")
         }
-        if (noteId != 0) editMode = true
+        if (noteId != 0L) editMode = true
     }
 
     override fun onCreateView(
@@ -73,7 +73,7 @@ class NoteDetailFragment : Fragment() {
     }
 
     private fun setUpViewModel() {
-        if (noteId != 0) {
+        if (noteId != 0L) {
             noteViewModel.getNote(noteId).observe(viewLifecycleOwner, {
                 note = it
                 binding.etTitle.setText(it.title)
